utils.PriorityQueue.java: priority queue with ordered list
This solves an issue with java's native PriorityQueue, in which the
order can't be guaranteed, so sometimes, when we needed to print a
queue, it would apparently be in the wrong order.
This implementation uses an ArrayList as container, and uses upperBound
algorithm to find in O(log n) time the index of insertion of new
elements.
Unfortunately, the whole insertion and removal operations are still
O(n), because the elements must be shifted left/right. A better approach
would be using trees, but currently there's no time to implement one.